# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for culinary recipes (BMSTU Team development project)",
        "title": "TeamDev API",
        "contact": {},
        "version": "0.1"
    },
    "paths": {
...
        "/recipes/{id}/steps": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steps"
                ],
                "summary": "Retrieves all recipe's steps",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/objects.StepDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Invalid"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steps"
                ],
                "summary": "Posts step",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Step",
                        "name": "step",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.StepDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/objects.StepDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Invalid"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "Authentication"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "Access"
                        }
                    }
                }
            }
        },
        "/recipes/{id}/steps/{step}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steps"
                ],
                "summary": "Retrieves recipe's step by its number",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Step's number",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/objects.StepDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Invalid"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steps"
                ],
                "summary": "Removes step",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Step's number",
                        "name": "step",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Successful"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "Invalid"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "Authentication"
                        }
                    }
                }
            },
            "patch": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Steps"
                ],
                "summary": "Updates step",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Recipe's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Step's number",
                        "name": "step",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Step data",
                        "name": "value",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/objects.StepDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Successful"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "objects.StepDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "num": {
                    "type": "integer"
                },
                "recipe": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}